Miscellaneous fixes (#2969)
* Suppressed various compilation warnings
* Fixed CodeQL warning
* Fixed build warning on Android and Linux
* Fix build warning on Visual Studio
* Update target Android SDK version to 29 (for pjsua2 & app), to avoid gradle error: 'Google Play requires that apps target API level 29 or higher'.
* Add support for Camera2 API to ipjsua android app.
* Fix assertion on Win32 app (MSVC2005) because ICE transport info size is 160 bytes while the allocated buffer is just 144 bytes.
* Enable python unit test on windows for github actions.
* Avoid assertion in pjsua test caused by double calling of pjsua_conf_remove_port()

+++ pjsip/src/pjsip/sip_transport.c
@@ -1843,7 +1843,7 @@ PJ_DEF(unsigned) pjsip_tpmgr_get_transport_count(pjsip_tpmgr *mgr)
     itr = pj_hash_first(mgr->table, &itr_val);
     while (itr) {
 	transport *tp_entry = (transport *)pj_hash_this(mgr->table, itr);
-	nr_of_transports += pj_list_size(tp_entry);
+	nr_of_transports += (int)pj_list_size(tp_entry);
 	itr = pj_hash_next(mgr->table, itr);
     }
 
@@ -2075,8 +2075,8 @@ PJ_DEF(pj_ssize_t) pjsip_tpmgr_receive_packet( pjsip_tpmgr *mgr,
 				       pj_exception_id_name(err->except_code),
 				       (int)err->hname.slen, err->hname.ptr,
 				       err->line, err->col);
-		if (len >= (int)sizeof(buf)-tmp.slen) {
-		    len = (int)sizeof(buf)-tmp.slen;
+		if (len >= (int)sizeof(buf)-(int)tmp.slen) {
+		    len = (int)sizeof(buf)-(int)tmp.slen;
 		}
 		if (len > 0) {
 		    tmp.slen += len;
